+++ packages/sampler/stories/moonstone-stories/VideoPlayer.js
@@ -134,7 +134,7 @@ storiesOf('VideoPlayer')
 						rateButtonsDisabled={boolean('rateButtonsDisabled', false)}
 						loop={boolean('loop', true)}
 						muted={boolean('muted', true)}
-						no5wayJump={boolean('no5wayJump', false)}
+						no5WayJump={boolean('no5WayJump', false)}
 						noAutoPlay={boolean('noAutoPlay', false)}
 						noJumpButtons={boolean('noJumpButtons', false)}
 						noRateButtons={boolean('noRateButtons', false)}