+++ packages/moonstone/CHANGELOG.md
@@ -8,7 +8,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 - `moonstone/VideoPlayer` ability to seek when holding down the right and left keys. Sensitivity can be adjusted using throttling options `jumpDelay` and `initialJumpDelay`.
-- `moonstone/VideoPlayer` property `no5wayJump` to disable jumping done by 5-way
+- `moonstone/VideoPlayer` property `no5WayJump` to disable jumping done by 5-way
 
 ### Changed